+++ pcs/cli/common/parse_args.py
@@ -70,7 +70,7 @@ def split_list_by_any_keywords(arg_list, keyword_label):
             groups[current_keyword] = []
     return groups
 
-def split_option(arg):
+def split_option(arg, allow_empty_value=True):
     """
     Get (key, value) from a key=value commandline argument.
 
@@ -78,6 +78,8 @@ def split_option(arg):
     CmdLineInputError if the argument cannot be splitted.
 
     string arg -- commandline argument
+    allow_empty_value -- if True, empty value is allowed. Otherwise,
+        CmdLineInputError exception is raised
 
     Commandline options: no options
     """
@@ -85,7 +87,10 @@ def split_option(arg):
         raise CmdLineInputError("missing value of '{0}' option".format(arg))
     if arg.startswith("="):
         raise CmdLineInputError("missing key in '{0}' option".format(arg))
-    return arg.split("=", 1)
+    key, value = arg.split("=", 1)
+    if not (value or allow_empty_value):
+        raise CmdLineInputError("value of '{0}' option is empty".format(key))
+    return key, value
 
 def prepare_options(cmdline_args, allowed_repeatable_options=()):
     """