+++ CHANGELOG.md
@@ -26,6 +26,8 @@
 - It is now possible to authenticate remote clusters in web UI even if the local
   cluster is not authenticated ([rhbz#1743735])
 - Documentation of `pcs constraint colocation add` ([rhbz#1734361])
+- Empty constraint option are not allowed in `pcs constraint order` and `pcs
+  constraint colocation add` commands ([rhbz#1734361])
 
 [rhbz#1442116]: https://bugzilla.redhat.com/show_bug.cgi?id=1442116
 [rhbz#1631514]: https://bugzilla.redhat.com/show_bug.cgi?id=1631514